+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-15T12:22:59.529Z",
+    "2026-07-15T13:36:20.309Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {