+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4042,7 +4042,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KXJVPS6Q9EAE06CAJ6WKE5NY"),
+  id: z.string().default("01KXJZWXBZKJTEVN7130GENDHT"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),