+++ packages/orq-rc/RELEASES.md
@@ -10849,3 +10849,13 @@ Based on:
 - [typescript v4.12.0-rc.30] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.30 - packages/orq-rc
+
+## 2026-07-15 13:39:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.31] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.31] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.31 - packages/orq-rc